unsigned int to size_t

+++ libft/submit/ft_memcpy.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void    *ft_memcpy(void *dest, const void *src, unsigned int n)
+void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
     unsigned char *d;
     unsigned char *s;
