+++ libft/submit/ft_strncmp.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+    size_t  i;
 
     i = 0;
     if (n == 0)
